move test helper method into separate static class
